v1.1.1 Nullable TemplateId, some values back to int

+++ AtomicAssetsClient/Data/TemplateStats.cs
@@ -2,8 +2,8 @@
 {
     public class TemplateStats
     {
-        public long Assets { get; set; }
+        public int Assets { get; set; }
 
-        public long Burned { get; set; }
+        public int Burned { get; set; }
     }
 }
